requirements.txt
numpy<2
opencv-contrib-python
pillow
pandas
streamlit
